requirements.txt
open-cv
websocket-clint
pandas
scikit-learn
pynput